refactor: Move Tyrian client package to `example`
